<refactor> 파일명 변경 및 텍스트스타일 추가
