+++ lib/screen/choose_floor_screen.dart
@@ -1,6 +1,6 @@
 /*층 선택 화면*/
 import 'package:flutter/material.dart';
-import 'package:flutter_app/components/textStyles.dart';
+import 'package:flutter_app/components/text_style.dart';
 import 'package:flutter_app/widget/bar.dart';
 import 'package:flutter_app/widget/floor_button.dart';
