+++ lib/screen/login_screen.dart
@@ -3,7 +3,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter_app/components/decorations.dart';
-import 'package:flutter_app/components/textStyles.dart';
+import 'package:flutter_app/components/text_style.dart';
 import 'package:flutter_app/widget/kakao_login_widget.dart';
 
 class LoginScreen extends StatefulWidget {
