+++ lib/widget/floor_button.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_app/components/textStyles.dart';
+import 'package:flutter_app/components/text_style.dart';
 import 'package:flutter_app/screen/robot_map_screen.dart';
 
 class FloorButton extends StatelessWidget {
